+++ frontend/src/components/Task/displaytext.tsx
@@ -13,15 +13,13 @@ export const DisplayText: React.VFC<{
   return (
     <foreignObject x={x} y={y} width={width} height={height} fill="none">
       <Box
-        style={{
-          width: '100%',
-          height: '100%',
-          textAlign: 'center',
-          padding: '4px',
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'center',
-        }}
+        w="full"
+        h="full"
+        p="4px"
+        textAlign="center"
+        display="flex"
+        alignItems="center"
+        justifyContent="center"
       >
         <span>{task.name}</span>
       </Box>
