+++ frontend/src/components/Task/label.tsx
@@ -2,6 +2,7 @@ import { Box } from '@chakra-ui/react';
 import { observer } from 'mobx-react-lite';
 
 import { TaskType } from '@/lib/models/Task';
+import { darkenColor } from '@/utils/darkenColor';
 
 export const Label: React.VFC<{
   task: TaskType;
@@ -26,23 +27,12 @@ export const Label: React.VFC<{
     x: labelSize.x / 2 - textAreaSize.x / 2,
     y: labelSize.y / 2 - textAreaSize.y / 2,
   };
-
-  const autoBorderColor = () => {
+  const darkenBorderColor = () => {
     if (typeof borderColor !== 'undefined') {
       return borderColor;
+    } else {
+      return darkenColor(color);
     }
-    const colorCode = color.slice(1, 7).match(/.{2}/g);
-    if (colorCode === null) {
-      return '#000000';
-    }
-    const borderColorCode = colorCode.map((c) => {
-      if (parseInt(c, 16) < 0x1e) {
-        return '00';
-      } else {
-        return (parseInt(c, 16) - 0x1e).toString(16).padStart(2, '0');
-      }
-    });
-    return '#' + borderColorCode.join('');
   };
 
   return (
@@ -50,7 +40,7 @@ export const Label: React.VFC<{
       <path
         d="M159.5 35C159.5 39.6671 157.34 44.1526 153.354 48.2761C149.365 52.4019 143.569 56.1403 136.368 59.2907C121.968 65.5908 102.039 69.5 80 69.5C57.9609 69.5 38.0321 65.5908 23.6319 59.2907C16.431 56.1403 10.635 52.4019 6.64627 48.2761C2.65984 44.1526 0.5 39.6671 0.5 35C0.5 30.3329 2.65984 25.8474 6.64627 21.724C10.635 17.5981 16.431 13.8597 23.6319 10.7093C38.0321 4.40922 57.9609 0.5 80 0.5C102.039 0.5 121.968 4.40922 136.368 10.7093C143.569 13.8597 149.365 17.5981 153.354 21.724C157.34 25.8474 159.5 30.3329 159.5 35Z"
         fill={color}
-        stroke={autoBorderColor()}
+        stroke={darkenBorderColor()}
       />
       <foreignObject
         x={textAreaPos.x}
@@ -60,13 +50,11 @@ export const Label: React.VFC<{
         fill="none"
       >
         <Box
-          style={{
-            width: '100%',
-            height: '100%',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-          }}
+          w="full"
+          h="full"
+          display="flex"
+          alignItems="center"
+          justifyContent="center"
         >
           <span>{task.label}</span>
         </Box>
